requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
pillow==10.4.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
